Remove redundendent functions in agent namespace
