+++ agent/plugins/common/agents_test.go
@@ -73,25 +73,6 @@ func TestLoadAgentRegistry_RejectsEmptyEntry(t *testing.T) {
 	assert.Contains(t, err.Error(), "must define globalDir and/or projectDir")
 }
 
-func TestParseHarnessList(t *testing.T) {
-	got, err := ParseHarnessList("cursor, Claude")
-	require.NoError(t, err)
-	assert.Equal(t, []string{"cursor", "claude"}, got)
-}
-
-func TestParseHarnessList_EmptyAndDuplicates(t *testing.T) {
-	_, err := ParseHarnessList("")
-	require.Error(t, err)
-
-	_, err = ParseHarnessList("cursor,,claude")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "empty name")
-
-	_, err = ParseHarnessList("cursor,cursor")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "more than once")
-}
-
 func TestResolveAgent_Unknown(t *testing.T) {
 	testutil.WithJfrogHome(t)
 	registry, err := agentcommon.LoadAgentRegistry(Agents, agentcommon.PluginsAgentsKey)
