+++ agent/skills/commands/install/install.go
@@ -247,13 +247,13 @@ func (ic *InstallCommand) resolveAgentTargetDirectories() ([]common.AgentTarget,
 		return ic.explicitTargets, nil
 	}
 	if ic.installPath != "" {
-		return common.ResolveAgentTargets(ic.slug, ic.installPath, nil, "", false)
+		return agentcommon.ResolveAgentTargets(ic.slug, ic.installPath, nil, "", false)
 	}
 	if ic.scope == agentcommon.InstallScopeProject && ic.projectDir == "" {
 		return nil, fmt.Errorf("project directory is required for project-scoped install")
 	}
 	isGlobal := ic.scope == agentcommon.InstallScopeGlobal
-	return common.ResolveAgentTargets(ic.slug, "", ic.agents, ic.projectDir, isGlobal)
+	return agentcommon.ResolveAgentTargets(ic.slug, "", ic.agents, ic.projectDir, isGlobal)
 }
 
 func (ic *InstallCommand) writeSkillInfoManifest(target common.AgentTarget) error {
