+++ agent/skills/common/agents_test.go
@@ -79,12 +79,6 @@ func TestLoadAgentRegistry_RejectsBadJSON(t *testing.T) {
 	assert.Contains(t, err.Error(), "failed to parse agent config")
 }
 
-func TestParseHarnessList(t *testing.T) {
-	got, err := ParseHarnessList("cursor, Claude-Code")
-	require.NoError(t, err)
-	assert.Equal(t, []string{"cursor", "claude-code"}, got)
-}
-
 func TestParseHarnessForList_Single(t *testing.T) {
 	got, err := ParseHarnessForList("cursor")
 	require.NoError(t, err)
@@ -95,19 +89,6 @@ func TestParseHarnessForList_RejectsCommaSeparated(t *testing.T) {
 	_, err := ParseHarnessForList("cursor,claude-code")
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "one harness name")
-}
-
-func TestParseHarnessList_EmptyAndDuplicates(t *testing.T) {
-	_, err := ParseHarnessList("")
-	require.Error(t, err)
-
-	_, err = ParseHarnessList("cursor,,claude-code")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "empty name")
-
-	_, err = ParseHarnessList("cursor,cursor")
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "more than once")
 }
 
 func TestResolveAgent_Unknown(t *testing.T) {
