+++ agent/skills/common/install_targets.go
@@ -2,7 +2,6 @@ package common
 
 import (
 	"fmt"
-	"path/filepath"
 	"strings"
 
 	agentcommon "github.com/jfrog/jfrog-cli-artifactory/agent/common"
@@ -49,7 +48,7 @@ func ValidateInstallFlags(c *components.Context) (agentcommon.InstallFlagsResult
 		return agentcommon.InstallFlagsResult{}, fmt.Errorf("--harness is required unless --path is set. Supported harnesses: %s", agentcommon.AgentNames(registry))
 	}
 
-	harnessNames, err := ParseHarnessList(rawHarness)
+	harnessNames, err := agentcommon.ParseHarnessList(rawHarness)
 	if err != nil {
 		return agentcommon.InstallFlagsResult{}, err
 	}
@@ -71,38 +70,4 @@ func ValidateInstallFlags(c *components.Context) (agentcommon.InstallFlagsResult
 		ProjectDirAbs: projectDirAbs,
 		IsGlobal:      isGlobal,
 	}, nil
-}
-
-// ResolveAgentTargets resolves per-agent install destinations.
-// When path is non-empty, a single ScopePath target is returned.
-func ResolveAgentTargets(slug, path string, agents []AgentSpec, projectDirAbs string, isGlobal bool) ([]AgentTarget, error) {
-	if path != "" {
-		target, err := agentcommon.BuildPathInstallTarget(slug, path)
-		if err != nil {
-			return nil, err
-		}
-		return []AgentTarget{target}, nil
-	}
-
-	scope := ScopeProject
-	if isGlobal {
-		scope = ScopeGlobal
-	}
-	if scope == ScopeProject && projectDirAbs == "" {
-		return nil, fmt.Errorf("project directory is required for project-scoped install")
-	}
-
-	targets := make([]AgentTarget, 0, len(agents))
-	for _, agent := range agents {
-		base, err := agentcommon.ResolveAgentInstallDir(agent, projectDirAbs, isGlobal)
-		if err != nil {
-			return nil, err
-		}
-		targets = append(targets, AgentTarget{
-			Agent:          agent,
-			Scope:          scope,
-			DestinationDir: filepath.Join(base, slug),
-		})
-	}
-	return targets, nil
 }
